Rename bzip back to bzip2, leave deprecated feature in
