+++ Cargo.toml
@@ -20,7 +20,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 default = []
 all = ["all-implementations", "all-algorithms"]
 all-implementations = ["bufread", "write", "stream"]
-all-algorithms = ["brotli", "bzip", "deflate", "gzip", "zlib", "zstd"]
+all-algorithms = ["brotli", "bzip2", "deflate", "gzip", "zlib", "zstd"]
 
 # implementations
 bufread = ["futures-io"]
@@ -29,7 +29,7 @@ stream = ["bytes"]
 
 # algorithms
 brotli = ["brotli2"]
-bzip = ["bzip2"]
+bzip = ["bzip2"]  # deprecated, please use `bzip2` instead
 deflate = ["flate2"]
 gzip = ["flate2"]
 zlib = ["flate2"]
@@ -61,8 +61,8 @@ name = "brotli"
 required-features = ["all-implementations", "brotli"]
 
 [[test]]
-name = "bzip"
-required-features = ["all-implementations", "bzip"]
+name = "bzip2"
+required-features = ["all-implementations", "bzip2"]
 
 [[test]]
 name = "deflate"
