+++ src/codec/bzip2/decoder.rs
