+++ src/codec/bzip2/encoder.rs
