+++ tests/bzip2.rs
@@ -1,4 +1,4 @@
 #[macro_use]
 mod utils;
 
-test_cases!(bzip);
+test_cases!(bzip2);
